[SIG-3110] Rename IncidentSplitContainer -> LegacyIncidentSplitContainer
In preparation of SIG-2473

(something went wrong with the previous pull request when I switched branches
so I've created this new pull request)
