--- src/signals/incident-management/containers/IncidentSplitContainer/constants.js
@@ -1,3 +1,0 @@
-export const SPLIT_INCIDENT = 'sia/IncidentSplitContainer/SPLIT_INCIDENT';
-export const SPLIT_INCIDENT_SUCCESS = 'sia/IncidentSplitContainer/SPLIT_INCIDENT_SUCCESS';
-export const SPLIT_INCIDENT_ERROR = 'sia/IncidentSplitContainer/SPLIT_INCIDENT_ERROR';
