+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/actions.js
