+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/components/IncidentPart/index.js
