+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/components/SplitDetail/index.test.js
