+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/constants.js
@@ -1,0 +1,3 @@
+export const SPLIT_INCIDENT = 'sia/LegacyIncidentSplitContainer/SPLIT_INCIDENT';
+export const SPLIT_INCIDENT_SUCCESS = 'sia/LegacyIncidentSplitContainer/SPLIT_INCIDENT_SUCCESS';
+export const SPLIT_INCIDENT_ERROR = 'sia/LegacyIncidentSplitContainer/SPLIT_INCIDENT_ERROR';
