+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/reducer.js
