+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/reducer.test.js
