+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/saga.js
