+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/selectors.test.js
