+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/services/formatUpdateIncident/index.js
