+++ src/signals/incident-management/containers/LegacyIncidentSplitContainer/services/useFetchIncident/index.js
